requirements.txt
streamlit==1.52.1
pandas==2.3.3
scikit-learn==1.8.0
requests==2.32.5
joblib==1.5.2
imbalanced-learn==0.14.0
python-dotenv==1.2.1
plotly==6.5.0
